requirements.txt
# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.0

# Evaluation and Metrics
evaluate>=0.4.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-watch>=4.2.0

# Visualization and Notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# CLI and Logging
rich>=13.7.0
typer>=0.9.0
loguru>=0.7.0

# Data Handling
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
datasets>=2.16.0  # For loading HuggingFace datasets
jsonlines>=3.1.0
faiss-cpu>=1.7.4

# Utilities
tqdm>=4.65.0

# Optional but recommended for large datasets
# vaex>=4.16.0
# pyarrow>=14.0.1